Gradient button componnet made
